+++ src/main/java/io/github/opencubicchunks/cubicchunks/core/asm/mixin/core/server/MixinDedicatedPlayerList_Bukkit.java
@@ -24,7 +24,6 @@
  */
 package io.github.opencubicchunks.cubicchunks.core.asm.mixin.core.server;
 
-import io.github.opencubicchunks.cubicchunks.core.asm.mixin.core.common.MixinPlayerList;
 import io.github.opencubicchunks.cubicchunks.core.asm.mixin.core.common.MixinPlayerList_Bukkit;
 import net.minecraft.server.dedicated.DedicatedPlayerList;
 import net.minecraft.server.dedicated.DedicatedServer;
