+++ src/main/resources/cubicchunks.mixins.core.bukkit.json
@@ -12,7 +12,7 @@
     "server": [
         "common.MixinWorld_HeightLimits_Bukkit_Sided",
         "common.MixinPlayerList_Bukkit",
-        "common.MixinChunk_Cubes_Bukkit",
+        "common.MixinChunk_Cubes_Bukkit_Sided",
         "server.MixinDedicatedPlayerList_Bukkit",
         "common.MixinWorldServer_Bukkit"
     ]
