+++ src/main/resources/cubicchunks.mixins.core.vanilla.json
@@ -8,7 +8,7 @@
         "conformVisibility": true
     },
     "mixins": [
-        "common.MixinChunk_Cubes",
+        "common.MixinChunk_Cubes_Vanilla_Sided",
         "common.MixinWorld_HeightLimits_Vanilla_Sided",
         "common.MixinWorldServer",
         "common.MixinPlayerList"
